refactor(issues): extract issue fetching and grouping logic into a composable
