+++ constants/columns.ts
@@ -1,0 +1,5 @@
+import { ISSUE_STATUSES } from './issues'
+
+export const COLUMN_TITLES = ['not-planned', ...ISSUE_STATUSES] as const
+
+export type ColumnTitle = (typeof COLUMN_TITLES)[number]
